+++ data/core/macros/ai.cfg
@@ -537,15 +537,6 @@
     # move-to-enemy candidate actions.
     # Put this into the [side][ai] tag.
     # Does not work in [modify_side][ai] or [modify_ai] at the moment.
-    [engine]
-        name="lua"
-        code= <<
-            local _,data = ...
-            local exp_ai = wesnoth.require("ai/lua/generic_rush_engine.lua").init(ai)
-            exp_ai.data = data
-            return exp_ai
-        >>
-    [/engine]
     [stage]
         id=main_loop
         name=ai_default_rca::candidate_action_evaluation_loop
@@ -564,57 +555,49 @@
             engine=lua
             name=recruit_rushers
             max_score=300000
-            evaluation="return (...):recruit_rushers_eval()"
-            execution="(...):recruit_rushers_exec()"
+            location="ai/lua/ca_recruit_rushers.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
             name=switch_castle
             max_score=290000
-            evaluation="return (...):castle_switch_eval()"
-            execution="(...):castle_switch_exec()"
+            location="ai/lua/ca_castle_switch.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
             name=retreat_injured
             max_score=205000
-            evaluation="return (...):retreat_injured_units_eval()"
-            execution="(...):retreat_injured_units_exec()"
+            location="ai/lua/ca_retreat_injured.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
             name=grab_villages
             max_score=200000
-            evaluation="return (...):grab_villages_eval()"
-            execution="(...):grab_villages_exec()"
+            location="ai/lua/ca_grab_villages.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
             name=spread_poison
             max_score=190000
-            evaluation="return (...):spread_poison_eval()"
-            execution="(...):spread_poison_exec()"
+            location="ai/lua/ca_spread_poison.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
             name=place_healers
             max_score=95000
-            evaluation="return (...):place_healers_eval()"
-            execution="(...):place_healers_exec()"
+            location="ai/lua/ca_place_healers.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
-            name=hunt_villages
+            name=village_hunt
             max_score=30000
-            evaluation="return (...):village_hunt_eval()"
-            execution="(...):village_hunt_exec()"
+            location="ai/lua/ca_village_hunt.lua"
         [/candidate_action]
         [candidate_action]
             engine=lua
-            name=move_to_enemy
+            name=move_to_any_enemy
             max_score=1
-            evaluation="return (...):move_to_enemy_eval()"
-            execution="(...):move_to_enemy_exec()"
+            location="ai/lua/ca_move_to_any_enemy.lua"
         [/candidate_action]
     [/stage]
 #enddef
